Reorganize the codebase to match the schedule-board redesign, fix bugs found in testing
The redesign (schedule board, custom theme, macOS .app bundle) grew the
codebase without a matching reorganization: main.go had become a 958-line
grab-bag of unrelated concerns, two test files duplicated coverage of the
same functions, and a few things were dead leftovers from the old list UI.

- Extract internal/config, internal/discovery, internal/platform packages
  for the pieces with no dependency on App/GameManagerUI; split the rest of
  main.go into config.go/launch.go/tray.go/logging.go by responsibility.
- Remove dead code (getGameScheduleInfo, splitStatusLabel, unused clockText
  field) and fold the one-function dialogs.go into its only caller.
- Consolidate main_test.go + coverage_test.go into one test file per source
  file, deduping redundant tests along the way.
- Drop config.example.yaml now that the in-app editor covers that need.

Dogfooding the reorganized build surfaced three real bugs, fixed alongside:
- A board block for a game with multiple schedule windows always showed the
  game's first window's time, regardless of which window that block was.
- The tray's "next up" labels were only recomputed on discrete events (save,
  reload, launch) and went stale once real time passed a shown window.
- A launch recorded in one schedule window on a given day incorrectly
  suppressed a different, later window for the same game that same day.
